grpc-demo/requirements.txt
grpcio-1.51.1
protobuf-4.21.12
grpcio-tools-1.51.1
googleapis-common-protos-1.58.0
astunparse-1.6.3 
protoc-gen-validate-0.9.1 
validate-email-1.3 
wheel-0.38.4